Refactor multiartifactlatefusion experiment code (#139)
**Motivation**: `train.py` was quite long. 

**Solution:** In order to alleviate this problem, code is put into other modules.

**Changes:**
* Move create_samples() function into preprocessing.py
* Move get_base_model() and download_pretrained_model() into model.py
* Put callbacks into utils() function
* Remove code for activation because unused
