+++ src/models/CNNDepthMap/CNNDepthMap-height/q3-depthmapmultiartifactlatefusion-plaincnn-height/src/constants.py
@@ -2,3 +2,5 @@ from pathlib import Path
 
 REPO_DIR = Path(__file__).parents[6].absolute()
 DATA_DIR_ONLINE_RUN = Path("/tmp/data/")
+MODEL_H5_FILENAME = "best_model.h5"
+MODEL_CKPT_FILENAME = "best_model.ckpt"
